Implement WebWire protocol v1.1
Implement the WebWire protocol v1.1 transitioning to a fully binary
protocol with built-in support for UTF8 and UTF16 encoding to support
both the default Golang and default JavaScript encoding as well as
plain binary data transmission.
Introduce named requests and signals.
Add tests to prove correct functioning of the new binary protocol.