Update WebWire protocol version to 1.3
In the fourth protocol revision the error reply message
moved from JSON to a binary encoding packing the error code
and message into separate fields.
A code length flag is included to define the size of the
ASCII 7 bit encoded error code field and message offset.
The UTF8 encoded message field is appended at the end of the message.