Fix parser behavior on requests
Make the webwire server accept requests that provide either
only a name or only a payload or both, but not none of them.
Make the webwire server not break the connection on parser errors
because protocol errors are not critical.
Add test cases to test requests with different
name/payload missing/provided combinations.