Introduce message & payload interface types
Introduce new message and payload interface types
which are now used by library hooks.
It is now possible to return nil for empty payloads.
All request-type methods of the client will now return a nil-payload
for failed calls when the returned error is not nil.
Refactor code by splitting the message parser
and constructor sources into a separate package.
Adjust tests and examples to the changes in the API.
Update README to reflect the current state of the API.