Improve session lookup hook interface usability
Improve the usability of the SessionManager's `OnSessionLookup` hook
by allowing to return `nil, nil` in case no session was found
(instead of enforcing the return of a `SessNotFoundErr` error)
and enforcing all mandatory session object fields
through the parameters of the new
`wwr.NewSessionLookupResult` function.
Remove `wwr.SessionFile` from the exported library namespace.
Update tests according to the changes in the library's API.