Transite to client-side heartbeating
Remove server-side heartbeating capabilities and add support for
client-side heartbeating instead by setting the read timeout on the
socket and reset it when receiving any kind of message including
a new heartbeat-type message.
Add heartbeating support to the client implementation.
Increment protocol version to 1.5.
Turn the ConnectionOptions interface into an actual struct
for increased readability and slightly increased performance.
Remove the "Heartbeat", "HeartbeatTimeout" and "HeartbeatInterval"
server options and add the new "ReadTimeout" option.
Set the default socket read timeout to 60 seconds.
Update examples and tests according to the changes in the API.
Add commands ":disconnect" and ":connect" to the chatroom example
client.